Add partial decoders and definitions -125ms startup -75MB ram
