+++ cache/src/main/kotlin/world/gregs/voidps/cache/definition/data/InterfaceComponentDefinitionFull.kt
@@ -3,105 +3,105 @@ package world.gregs.voidps.cache.definition.data
 import world.gregs.voidps.cache.Definition
 import world.gregs.voidps.cache.definition.Extra
 
-internal data class InterfaceComponentDefinitionFull(
+data class InterfaceComponentDefinitionFull(
     override var id: Int = -1,
-    internal var type: Int = 0,
-    internal var unknown: String? = null,
-    internal var contentType: Int = 0,
-    internal var basePositionX: Int = 0,
-    internal var basePositionY: Int = 0,
-    internal var baseWidth: Int = 0,
-    internal var baseHeight: Int = 0,
-    internal var horizontalSizeMode: Byte = 0,
-    internal var verticalSizeMode: Byte = 0,
-    internal var horizontalPositionMode: Byte = 0,
-    internal var verticalPositionMode: Byte = 0,
-    internal var parent: Int = -1,
-    internal var hidden: Boolean = false,
-    internal var disableHover: Boolean = false,
-    internal var scrollWidth: Int = 0,
-    internal var scrollHeight: Int = 0,
-    internal var colour: Int = 0,
-    internal var filled: Boolean = false,
-    internal var alpha: Int = 0,
-    internal var fontId: Int = -1,
-    internal var monochrome: Boolean = true,
-    internal var text: String = "",
-    internal var lineHeight: Int = 0,
-    internal var horizontalTextAlign: Int = 0,
-    internal var verticalTextAlign: Int = 0,
-    internal var shaded: Boolean = false,
-    internal var lineCount: Int = 0,
-    internal var defaultImage: Int = -1,
-    internal var imageRotation: Int = 0,
-    internal var aBoolean4861: Boolean = false,
-    internal var imageRepeat: Boolean = false,
-    internal var rotation: Int = 0,
-    internal var backgroundColour: Int = 0,
-    internal var flipVertical: Boolean = false,
-    internal var flipHorizontal: Boolean = false,
-    internal var aBoolean4782: Boolean = true,
-    internal var defaultMediaType: Int = 1,
-    internal var defaultMediaId: Int = 0,
-    internal var animated: Boolean = false,
-    internal var centreType: Boolean = false,
-    internal var ignoreZBuffer: Boolean = false,
-    internal var viewportX: Int = 0,
-    internal var viewportY: Int = 0,
-    internal var viewportZ: Int = 0,
-    internal var spritePitch: Int = 0,
-    internal var spriteRoll: Int = 0,
-    internal var spriteYaw: Int = 0,
-    internal var spriteScale: Int = 100,
-    internal var animation: Int = -1,
-    internal var viewportWidth: Int = 0,
-    internal var viewportHeight: Int = 0,
-    internal var lineWidth: Int = 1,
-    internal var lineMirrored: Boolean = false,
-    internal var keyRepeat: ByteArray? = null,
-    internal var keyCodes: ByteArray? = null,
-    internal var keyModifiers: IntArray? = null,
-    internal var clickable: Boolean = false,
-    internal var name: String = "",
+    var type: Int = 0,
+    var unknown: String? = null,
+    var contentType: Int = 0,
+    var basePositionX: Int = 0,
+    var basePositionY: Int = 0,
+    var baseWidth: Int = 0,
+    var baseHeight: Int = 0,
+    var horizontalSizeMode: Byte = 0,
+    var verticalSizeMode: Byte = 0,
+    var horizontalPositionMode: Byte = 0,
+    var verticalPositionMode: Byte = 0,
+    var parent: Int = -1,
+    var hidden: Boolean = false,
+    var disableHover: Boolean = false,
+    var scrollWidth: Int = 0,
+    var scrollHeight: Int = 0,
+    var colour: Int = 0,
+    var filled: Boolean = false,
+    var alpha: Int = 0,
+    var fontId: Int = -1,
+    var monochrome: Boolean = true,
+    var text: String = "",
+    var lineHeight: Int = 0,
+    var horizontalTextAlign: Int = 0,
+    var verticalTextAlign: Int = 0,
+    var shaded: Boolean = false,
+    var lineCount: Int = 0,
+    var defaultImage: Int = -1,
+    var imageRotation: Int = 0,
+    var aBoolean4861: Boolean = false,
+    var imageRepeat: Boolean = false,
+    var rotation: Int = 0,
+    var backgroundColour: Int = 0,
+    var flipVertical: Boolean = false,
+    var flipHorizontal: Boolean = false,
+    var aBoolean4782: Boolean = true,
+    var defaultMediaType: Int = 1,
+    var defaultMediaId: Int = 0,
+    var animated: Boolean = false,
+    var centreType: Boolean = false,
+    var ignoreZBuffer: Boolean = false,
+    var viewportX: Int = 0,
+    var viewportY: Int = 0,
+    var viewportZ: Int = 0,
+    var spritePitch: Int = 0,
+    var spriteRoll: Int = 0,
+    var spriteYaw: Int = 0,
+    var spriteScale: Int = 100,
+    var animation: Int = -1,
+    var viewportWidth: Int = 0,
+    var viewportHeight: Int = 0,
+    var lineWidth: Int = 1,
+    var lineMirrored: Boolean = false,
+    var keyRepeat: ByteArray? = null,
+    var keyCodes: ByteArray? = null,
+    var keyModifiers: IntArray? = null,
+    var clickable: Boolean = false,
+    var name: String = "",
     var options: Array<String>? = null,
-    internal var mouseIcon: IntArray? = null,
-    internal var optionOverride: String? = null,
-    internal var anInt4708: Int = 0,// Drag type
-    internal var anInt4795: Int = 0,// Drag slider?
-    internal var anInt4860: Int = 0,// Friends list icons/buttons?
-    internal var useOption: String = "",
-    internal var anInt4698: Int = -1,
-    internal var anInt4839: Int = -1,// Unused
-    internal var anInt4761: Int = -1,// Unused
-    internal var setting: InterfaceComponentSetting = InterfaceComponentSetting(0, -1),
-    internal val params: HashMap<Long, Any>? = null,
+    var mouseIcon: IntArray? = null,
+    var optionOverride: String? = null,
+    var anInt4708: Int = 0,// Drag type
+    var anInt4795: Int = 0,// Drag slider?
+    var anInt4860: Int = 0,// Friends list icons/buttons?
+    var useOption: String = "",
+    var anInt4698: Int = -1,
+    var anInt4839: Int = -1,// Unused
+    var anInt4761: Int = -1,// Unused
+    var setting: InterfaceComponentSetting = InterfaceComponentSetting(0, -1),
+    val params: HashMap<Long, Any>? = null,
     var anObjectArray4758: Array<Any>? = null,
-    internal var mouseEnterHandler: Array<Any>? = null,
-    internal var mouseExitHandler: Array<Any>? = null,
-    internal var anObjectArray4771: Array<Any>? = null,
-    internal var anObjectArray4768: Array<Any>? = null,
-    internal var stateChangeHandler: Array<Any>? = null,
-    internal var invUpdateHandler: Array<Any>? = null,
-    internal var refreshHandler: Array<Any>? = null,
-    internal var updateHandler: Array<Any>? = null,
-    internal var anObjectArray4770: Array<Any>? = null,
-    internal var anObjectArray4751: Array<Any>? = null,
-    internal var mouseMotionHandler: Array<Any>? = null,
-    internal var mousePressedHandler: Array<Any>? = null,
-    internal var mouseDraggedHandler: Array<Any>? = null,
-    internal var mouseReleasedHandler: Array<Any>? = null,
-    internal var mouseDragPassHandler: Array<Any>? = null,
-    internal var anObjectArray4852: Array<Any>? = null,
-    internal var anObjectArray4711: Array<Any>? = null,
-    internal var anObjectArray4753: Array<Any>? = null,
-    internal var anObjectArray4688: Array<Any>? = null,
-    internal var anObjectArray4775: Array<Any>? = null,
-    internal var clientVarp: IntArray? = null,
-    internal var containers: IntArray? = null,
-    internal var anIntArray4789: IntArray? = null,
-    internal var clientVarc: IntArray? = null,
-    internal var anIntArray4805: IntArray? = null,
-    internal var hasScript: Boolean = false,
+    var mouseEnterHandler: Array<Any>? = null,
+    var mouseExitHandler: Array<Any>? = null,
+    var anObjectArray4771: Array<Any>? = null,
+    var anObjectArray4768: Array<Any>? = null,
+    var stateChangeHandler: Array<Any>? = null,
+    var invUpdateHandler: Array<Any>? = null,
+    var refreshHandler: Array<Any>? = null,
+    var updateHandler: Array<Any>? = null,
+    var anObjectArray4770: Array<Any>? = null,
+    var anObjectArray4751: Array<Any>? = null,
+    var mouseMotionHandler: Array<Any>? = null,
+    var mousePressedHandler: Array<Any>? = null,
+    var mouseDraggedHandler: Array<Any>? = null,
+    var mouseReleasedHandler: Array<Any>? = null,
+    var mouseDragPassHandler: Array<Any>? = null,
+    var anObjectArray4852: Array<Any>? = null,
+    var anObjectArray4711: Array<Any>? = null,
+    var anObjectArray4753: Array<Any>? = null,
+    var anObjectArray4688: Array<Any>? = null,
+    var anObjectArray4775: Array<Any>? = null,
+    var clientVarp: IntArray? = null,
+    var containers: IntArray? = null,
+    var anIntArray4789: IntArray? = null,
+    var clientVarc: IntArray? = null,
+    var anIntArray4805: IntArray? = null,
+    var hasScript: Boolean = false,
     override var stringId: String = "",
     override var extras: Map<String, Any>? = null
 ) : Definition, Extra {
