+++ cache/src/main/kotlin/world/gregs/voidps/cache/definition/decoder/InterfaceDecoder.kt
@@ -7,8 +7,6 @@ import world.gregs.voidps.cache.Cache
 import world.gregs.voidps.cache.DefinitionDecoder
 import world.gregs.voidps.cache.Index.INTERFACES
 import world.gregs.voidps.cache.definition.data.InterfaceComponentDefinition
-import world.gregs.voidps.cache.definition.data.InterfaceComponentDefinitionFull
-import world.gregs.voidps.cache.definition.data.InterfaceComponentSetting
 import world.gregs.voidps.cache.definition.data.InterfaceDefinition
 
 class InterfaceDecoder : DefinitionDecoder<InterfaceDefinition>(INTERFACES) {
@@ -25,9 +23,9 @@ class InterfaceDecoder : DefinitionDecoder<InterfaceDefinition>(INTERFACES) {
         if (definition.components == null) {
             definition.components = Int2ObjectOpenHashMap()
         }
-        val component = InterfaceComponentDefinitionFull(id = id)
+        val component = InterfaceComponentDefinition(id = id)
         component.read(reader)
-        definition.components!![InterfaceDefinition.componentId(id)] = component.toLite()
+        definition.components!![InterfaceDefinition.componentId(id)] = component
     }
 
     override fun load(definitions: Array<InterfaceDefinition>, cache: Cache, id: Int) {
@@ -39,201 +37,111 @@ class InterfaceDecoder : DefinitionDecoder<InterfaceDefinition>(INTERFACES) {
         val definition = definitions[id]
         val components = Int2ObjectOpenHashMap<InterfaceComponentDefinition>(lastArchive)
         for (file in 0..lastArchive) {
-            val component = InterfaceComponentDefinitionFull(id = file + (id shl 16))
+            val component = InterfaceComponentDefinition(id = file + (id shl 16))
             val data = cache.getFile(index, archiveId, file)
             if (data != null) {
                 component.read(BufferReader(data))
             }
-            components[file] = component.toLite()
+            components[file] = component
         }
         definition.components = components
     }
 
-    internal fun InterfaceComponentDefinitionFull.read(buffer: Reader) {
+    private fun InterfaceComponentDefinition.read(buffer: Reader) {
         buffer.readUnsignedByte()
-        type = buffer.readUnsignedByte()
+        var type = buffer.readUnsignedByte()
         if (type and 0x80 != 0) {
             type = type and 0x7f
-            unknown = buffer.readString()
+            buffer.readString()
         }
-        contentType = buffer.readShort()
-        basePositionX = buffer.readUnsignedShort()
-        basePositionY = buffer.readUnsignedShort()
-        baseWidth = buffer.readShort()
-        baseHeight = buffer.readShort()
-        horizontalSizeMode = buffer.readByte().toByte()
-        verticalSizeMode = buffer.readByte().toByte()
-        horizontalPositionMode = buffer.readByte().toByte()
-        verticalPositionMode = buffer.readByte().toByte()
-        parent = buffer.readShort()
-        if (parent == 65535) {
-            parent = -1
-        }
-        val i_17_ = buffer.readUnsignedByte()
-        hidden = 0x1 and i_17_ != 0
+        buffer.skip(10)
+        val horizontalSizeMode = buffer.readByte()
+        val verticalSizeMode = buffer.readByte()
+        buffer.skip(5)
         if (type == 0) {
-            scrollWidth = buffer.readShort()
-            scrollHeight = buffer.readShort()
-            disableHover = buffer.readUnsignedByte() == 1
+            buffer.readShort()
+            buffer.readShort()
+            buffer.skip(1)
         } else if (type == 3) {
-            colour = buffer.readInt()
-            filled = buffer.readUnsignedByte() == 1
-            alpha = buffer.readUnsignedByte()
+            buffer.skip(6)
         } else if (type == 4) {
-            fontId = buffer.readShort()
-            if (fontId == 65535) {
-                fontId = -1
-            }
-            text = buffer.readString()
-            lineHeight = buffer.readUnsignedByte()
-            horizontalTextAlign = buffer.readUnsignedByte()
-            verticalTextAlign = buffer.readUnsignedByte()
-            shaded = buffer.readUnsignedByte() == 1
-            colour = buffer.readInt()
-            alpha = buffer.readUnsignedByte()
+            buffer.skip(2)
+            buffer.readString()
+            buffer.skip(9)
         } else if (type == 5) {
-            defaultImage = buffer.readInt()
-            imageRotation = buffer.readShort()
-            val i_18_ = buffer.readUnsignedByte()
-            aBoolean4861 = i_18_ and 0x1 != 0
-            imageRepeat = i_18_ and 0x2 != 0
-            alpha = buffer.readUnsignedByte()
-            rotation = buffer.readUnsignedByte()
-            backgroundColour = buffer.readInt()
-            flipVertical = buffer.readUnsignedByte() == 1
-            flipHorizontal = buffer.readUnsignedByte() == 1
-            colour = buffer.readInt()
+            buffer.skip(19)
         } else if (type == 6) {
-            defaultMediaType = 1
-            defaultMediaId = buffer.readShort()
-            if (defaultMediaId == 65535) {
-                defaultMediaId = -1
-            }
-            val i_19_ = buffer.readUnsignedByte()
-            animated = 0x4 and i_19_ == 4
-            val bool = 0x1 and i_19_ == 1
-            centreType = i_19_ and 0x2 == 2
-            ignoreZBuffer = 0x8 and i_19_ == 8
+            buffer.skip(2)
+            val data = buffer.readUnsignedByte()
+            val bool = 0x1 and data == 1
+            val centreType = data and 0x2 == 2
             if (bool) {
-                viewportX = buffer.readUnsignedShort()
-                viewportY = buffer.readUnsignedShort()
-                spritePitch = buffer.readShort()
-                spriteRoll = buffer.readShort()
-                spriteYaw = buffer.readShort()
-                spriteScale = buffer.readShort()
+                buffer.skip(12)
             } else if (centreType) {
-                viewportX = buffer.readUnsignedShort()
-                viewportY = buffer.readUnsignedShort()
-                viewportZ = buffer.readUnsignedShort()
-                spritePitch = buffer.readShort()
-                spriteRoll = buffer.readShort()
-                spriteYaw = buffer.readShort()
-                spriteScale = buffer.readUnsignedShort()
+                buffer.skip(14)
             }
-            animation = buffer.readShort()
-            if (animation == 65535) {
-                animation = -1
+            buffer.skip(2)
+            if (horizontalSizeMode != 0) {
+                buffer.skip(2)
             }
-            if (horizontalSizeMode.toInt() != 0) {
-                viewportWidth = buffer.readShort()
-            }
-            if (verticalSizeMode.toInt() != 0) {
-                viewportHeight = buffer.readShort()
+            if (verticalSizeMode != 0) {
+                buffer.skip(2)
             }
         } else if (type == 9) {
-            lineWidth = buffer.readUnsignedByte()
-            colour = buffer.readInt()
-            lineMirrored = buffer.readUnsignedByte() == 1
+            buffer.skip(6)
         }
         val setting = buffer.readUnsignedMedium()
-        var i_21_ = buffer.readUnsignedByte()
-        if (i_21_ != 0) {
-            keyRepeat = ByteArray(11)
-            keyCodes = ByteArray(11)
-            keyModifiers = IntArray(11)
-            while (i_21_ != 0) {
-                val i_22_ = (i_21_ shr 4) - 1
-                i_21_ = buffer.readUnsignedByte() or i_21_ shl 8
-                i_21_ = i_21_ and 0xfff
-                if (i_21_ == 4095) {
-                    i_21_ = -1
-                }
-                val b_23_ = buffer.readByte().toByte()
-                if (b_23_.toInt() != 0) {
-                    clickable = true
-                }
-                val b_24_ = buffer.readByte().toByte()
-                keyModifiers!![i_22_] = i_21_
-                keyRepeat!![i_22_] = b_23_
-                keyCodes!![i_22_] = b_24_
-                i_21_ = buffer.readUnsignedByte()
+        var code = buffer.readUnsignedByte()
+        if (code != 0) {
+            while (code != 0) {
+                buffer.skip(3)
+                code = buffer.readUnsignedByte()
             }
         }
-        name = buffer.readString()
-        val i_25_ = buffer.readUnsignedByte()
-        val optionCount = i_25_ and 0xf
+        buffer.readString()
+        val data = buffer.readUnsignedByte()
+        val optionCount = data and 0xf
         if (optionCount > 0) {
             options = Array(optionCount) { buffer.readString() }
         }
-        val iconCount = i_25_ shr 4
+        val iconCount = data shr 4
         if (iconCount > 0) {
-            val i_29_ = buffer.readUnsignedByte()
-            mouseIcon = IntArray(i_29_ + 1) { if (i_29_ + 1 > it) -1 else 0 }
-            mouseIcon!![i_29_] = buffer.readShort()
+            buffer.skip(3)
         }
         if (iconCount > 1) {
-            val i_31_ = buffer.readUnsignedByte()
-            mouseIcon!![i_31_] = buffer.readShort()
+            buffer.skip(3)
         }
-        optionOverride = buffer.readString()
-        if (optionOverride == "") {
-            optionOverride = null
-        }
-        anInt4708 = buffer.readUnsignedByte()
-        anInt4795 = buffer.readUnsignedByte()
-        anInt4860 = buffer.readUnsignedByte()
-        useOption = buffer.readString()
-        var i_32_ = -1
+        buffer.readString()
+        buffer.skip(3)
+        buffer.readString()
         if (setting and 0x3fda8 shr 11 != 0) {
-            i_32_ = buffer.readShort()
-            if (i_32_ == 65535) {
-                i_32_ = -1
-            }
-            anInt4698 = buffer.readShort()
-            if (anInt4698 == 65535) {
-                anInt4698 = -1
-            }
-            anInt4839 = buffer.readShort()
-            if (anInt4839 == 65535) {
-                anInt4839 = -1
-            }
+            buffer.skip(6)
         }
-        this.setting = InterfaceComponentSetting(setting, i_32_)
         anObjectArray4758 = decodeScript(buffer)
-        mouseEnterHandler = decodeScript(buffer)
-        mouseExitHandler = decodeScript(buffer)
-        anObjectArray4771 = decodeScript(buffer)
-        anObjectArray4768 = decodeScript(buffer)
-        stateChangeHandler = decodeScript(buffer)
-        invUpdateHandler = decodeScript(buffer)
-        refreshHandler = decodeScript(buffer)
-        updateHandler = decodeScript(buffer)
-        anObjectArray4770 = decodeScript(buffer)
-        mouseMotionHandler = decodeScript(buffer)
-        mousePressedHandler = decodeScript(buffer)
-        mouseDraggedHandler = decodeScript(buffer)
-        mouseReleasedHandler = decodeScript(buffer)
-        mouseDragPassHandler = decodeScript(buffer)
-        anObjectArray4852 = decodeScript(buffer)
-        anObjectArray4711 = decodeScript(buffer)
-        anObjectArray4753 = decodeScript(buffer)
-        anObjectArray4688 = decodeScript(buffer)
-        anObjectArray4775 = decodeScript(buffer)
-        clientVarp = decodeIntArray(buffer)
-        containers = decodeIntArray(buffer)
-        anIntArray4789 = decodeIntArray(buffer)
-        clientVarc = decodeIntArray(buffer)
-        anIntArray4805 = decodeIntArray(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        skipScript(buffer)
+        buffer.skip(buffer.readUnsignedByte() * 4)
+        buffer.skip(buffer.readUnsignedByte() * 4)
+        buffer.skip(buffer.readUnsignedByte() * 4)
+        buffer.skip(buffer.readUnsignedByte() * 4)
+        buffer.skip(buffer.readUnsignedByte() * 4)
     }
 
     override fun InterfaceDefinition.read(opcode: Int, buffer: Reader) {
@@ -241,34 +149,23 @@ class InterfaceDecoder : DefinitionDecoder<InterfaceDefinition>(INTERFACES) {
     }
 
     companion object {
-
-        private fun InterfaceComponentDefinitionFull.toLite() = InterfaceComponentDefinition(
-            id = id,
-            options = options,
-            anObjectArray4758 = anObjectArray4758,
-            stringId = stringId,
-            extras = extras,
-        )
-
-        private fun InterfaceComponentDefinitionFull.decodeScript(buffer: Reader): Array<Any>? {
+        private fun decodeScript(buffer: Reader): Array<Any>? {
             val length = buffer.readUnsignedByte()
             if (length == 0) {
                 return null
             }
-            val objects = Array<Any>(length) {
-                val string = buffer.readUnsignedBoolean()
-                if (string) buffer.readString() else buffer.readInt()
-            }
-            hasScript = true
-            return objects
+            return Array(length) { if (buffer.readUnsignedBoolean()) buffer.readString() else buffer.readInt() }
         }
 
-        private fun decodeIntArray(buffer: Reader): IntArray? {
+        private fun skipScript(buffer: Reader) {
             val length = buffer.readUnsignedByte()
             if (length == 0) {
-                return null
+                return
             }
-            return IntArray(length) { buffer.readInt() }
+            for (i in 0 until length) {
+                val string = buffer.readUnsignedBoolean()
+                if (string) buffer.readString() else buffer.skip(4)
+            }
         }
     }
 }
