+++ cache/src/main/kotlin/world/gregs/voidps/cache/definition/decoder/NPCDecoder.kt
@@ -15,62 +15,20 @@ class NPCDecoder(val member: Boolean) : DefinitionDecoder<NPCDefinition>(NPCS) {
 
     override fun NPCDefinition.read(opcode: Int, buffer: Reader) {
         when (opcode) {
-            1 -> {
-                val length = buffer.readUnsignedByte()
-                modelIds = IntArray(length)
-                for (count in 0 until length) {
-                    modelIds!![count] = buffer.readUnsignedShort()
-                    if (modelIds!![count] == 65535) {
-                        modelIds!![count] = -1
-                    }
-                }
-            }
+            1, 60, 160 -> buffer.skip(buffer.readUnsignedByte() * 2)
             2 -> name = buffer.readString()
             12 -> size = buffer.readUnsignedByte()
             in 30..34 -> options[-30 + opcode] = buffer.readString()
-            40 -> readColours(buffer)
-            41 -> readTextures(buffer)
-            42 -> recolourPalette = ByteArray(buffer.readUnsignedByte()) { buffer.readByte().toByte() }
-            60 -> dialogueModels = IntArray(buffer.readUnsignedByte()) { buffer.readUnsignedShort() }
-            93 -> drawMinimapDot = false
+            40, 41, 121 -> buffer.skip(buffer.readUnsignedByte() * 4)
+            42 -> buffer.skip(buffer.readUnsignedByte())
+            93, 99, 107, 109, 111, 141, 143, 158, 159, 162 -> return
             95 -> combat = buffer.readShort()
-            97 -> scaleXY = buffer.readShort()
-            98 -> scaleZ = buffer.readShort()
-            99 -> priorityRender = true
-            100 -> lightModifier = buffer.readByte()
-            101 -> shadowModifier = 5 * buffer.readByte()
-            102 -> headIcon = buffer.readShort()
-            103 -> rotation = buffer.readShort()
+            97, 98, 102, 103, 114, 122, 123, 137, 138, 139, 142 -> buffer.skip(2)
+            100, 101, 125, 128, 140, 163, 165, 168 -> buffer.skip(1)
             106, 118 -> readTransforms(buffer, opcode == 118)
-            107 -> clickable = false
-            109 -> slowWalk = false
-            111 -> animateIdle = false
-            113 -> {
-                primaryShadowColour = buffer.readShort().toShort()
-                secondaryShadowColour = buffer.readShort().toShort()
-            }
-            114 -> {
-                primaryShadowModifier = buffer.readByte().toByte()
-                secondaryShadowModifier = buffer.readByte().toByte()
-            }
+            113, 155, 164 -> buffer.skip(4)
             119 -> walkMask = buffer.readByte().toByte()
-            121 -> {
-                translations = arrayOfNulls(modelIds!!.size)
-                val length = buffer.readUnsignedByte()
-                for (count in 0 until length) {
-                    val index = buffer.readUnsignedByte()
-                    translations!![index] = intArrayOf(
-                        buffer.readByte(),
-                        buffer.readByte(),
-                        buffer.readByte()
-                    )
-                }
-            }
-            122 -> hitbarSprite = buffer.readShort()
-            123 -> height = buffer.readShort()
-            125 -> respawnDirection = buffer.readByte().toByte()
             127 -> renderEmote = buffer.readShort()
-            128 -> buffer.readUnsignedByte()
             134 -> {
                 idleSound = buffer.readShort()
                 if (idleSound == 65535) {
@@ -90,47 +48,13 @@ class NPCDecoder(val member: Boolean) : DefinitionDecoder<NPCDefinition>(NPCS) {
                 }
                 soundDistance = buffer.readUnsignedByte()
             }
-            135 -> {
-                primaryCursorOp = buffer.readUnsignedByte()
-                primaryCursor = buffer.readShort()
-            }
-            136 -> {
-                secondaryCursorOp = buffer.readUnsignedByte()
-                secondaryCursor = buffer.readShort()
-            }
-            137 -> attackCursor = buffer.readShort()
-            138 -> armyIcon = buffer.readShort()
-            139 -> spriteId = buffer.readShort()
-            140 -> ambientSoundVolume = buffer.readUnsignedByte()
-            141 -> visiblePriority = true
-            142 -> mapFunction = buffer.readShort()
-            143 -> invisiblePriority = true
+            135, 136 -> buffer.skip(3)
             in 150..154 -> {
                 options[opcode - 150] = buffer.readString()
                 if (!member) {
                     options[opcode - 150] = null
                 }
             }
-            155 -> {
-                hue = buffer.readByte().toByte()
-                saturation = buffer.readByte().toByte()
-                lightness = buffer.readByte().toByte()
-                opacity = buffer.readByte().toByte()
-            }
-            158 -> mainOptionIndex = 1.toByte()
-            159 -> mainOptionIndex = 0.toByte()
-            160 -> {
-                val length = buffer.readUnsignedByte()
-                campaigns = IntArray(length) { buffer.readShort() }
-            }
-            162 -> aBoolean2883 = true
-            163 -> anInt2803 = buffer.readUnsignedByte()
-            164 -> {
-                anInt2844 = buffer.readShort()
-                anInt2852 = buffer.readShort()
-            }
-            165 -> anInt2831 = buffer.readUnsignedByte()
-            168 -> anInt2862 = buffer.readUnsignedByte()
             249 -> readParameters(buffer)
         }
     }
