+++ cache/src/main/kotlin/world/gregs/voidps/cache/definition/decoder/ObjectDecoder.kt
@@ -23,11 +23,9 @@ open class ObjectDecoder(
                     skip(buffer)
                 }
                 val length = buffer.readUnsignedByte()
-                modelTypes = ByteArray(length)
-                this.modelIds = Array(length) { count ->
-                    modelTypes!![count] = buffer.readByte().toByte()
-                    val size = buffer.readUnsignedByte()
-                    IntArray(size) { buffer.readUnsignedShort() }
+                for (i in 0 until length) {
+                    buffer.skip(1)
+                    buffer.skip(buffer.readUnsignedByte() * 2)
                 }
                 if (opcode == 5 && !lowDetail) {
                     skip(buffer)
@@ -46,105 +44,28 @@ open class ObjectDecoder(
                 block = block and ObjectDefinition.PROJECTILE.inv()
             }
             19 -> interactive = buffer.readUnsignedByte()
-            21 -> contouredGround = 1
-            22 -> delayShading = true
-            23 -> culling = 1
-            24 -> {
-                val length = buffer.readShort()
-                if (length != 65535) {
-                    animations = intArrayOf(length)
-                }
-            }
+            21, 22, 23, 64, 73, 82, 88, 89, 91, 94, 97, 98, 103, 105, 168, 169, 177 -> return
+            24, 65, 66, 67, 70, 71, 72, 93, 95, 102, 107, 164, 165, 166, 167 -> buffer.skip(2)
             27 -> solid = 1
-            28 -> offsetMultiplier = buffer.readUnsignedByte() shl 2
-            29 -> brightness = buffer.readByte()
+            28, 29, 39, 75, 81, 101, 104, 178 -> buffer.skip(1)
             in 30..34 -> {
                 if (options == null) {
                     options = arrayOf(null, null, null, null, null, "Examine")
                 }
                 options!![opcode - 30] = buffer.readString()
             }
-            39 -> contrast = buffer.readByte() * 5
-            40 -> readColours(buffer)
-            41 -> readTextures(buffer)
-            42 -> readColourPalette(buffer)
+            40, 41 -> buffer.skip(buffer.readUnsignedByte() * 4)
+            42 -> buffer.skip(buffer.readUnsignedByte())
             62 -> mirrored = true
-            64 -> castsShadow = false
-            65 -> modelSizeX = buffer.readShort()
-            66 -> modelSizeZ = buffer.readShort()
-            67 -> modelSizeY = buffer.readShort()
             69 -> blockFlag = buffer.readUnsignedByte()
-            70 -> offsetX = buffer.readUnsignedShort() shl 2
-            71 -> offsetZ = buffer.readUnsignedShort() shl 2
-            72 -> offsetY = buffer.readUnsignedShort() shl 2
-            73 -> blocksLand = true
-            74 -> {
-                ignoreOnRoute = true
-                block = block and ObjectDefinition.ROUTE.inv()
-            }
-            75 -> supportItems = buffer.readUnsignedByte()
+            74 -> block = block and ObjectDefinition.ROUTE.inv()
             77, 92 -> readTransforms(buffer, opcode == 92)
-            78 -> {
-                anInt3015 = buffer.readShort()
-                anInt3012 = buffer.readUnsignedByte()
-            }
+            78, 99, 100 -> buffer.skip(3)
             79 -> {
-                anInt2989 = buffer.readShort()
-                anInt2971 = buffer.readShort()
-                anInt3012 = buffer.readUnsignedByte()
-                val length = buffer.readUnsignedByte()
-                anIntArray3036 = IntArray(length) { buffer.readShort() }
+                buffer.skip(5)
+                buffer.skip(buffer.readUnsignedByte() * 2)
             }
-            81 -> {
-                contouredGround = 2.toByte()
-                anInt3023 = buffer.readUnsignedByte() * 256
-            }
-            82 -> hideMinimap = true
-            88 -> aBoolean2972 = false
-            89 -> animateImmediately = false
-            91 -> isMembers = true
-            93 -> {
-                contouredGround = 3
-                anInt3023 = buffer.readShort()
-            }
-            94 -> contouredGround = 4
-            95 -> {
-                contouredGround = 5
-                anInt3023 = buffer.readUnsignedShort()
-            }
-            97 -> aBoolean3056 = true
-            98 -> aBoolean2998 = true
-            99 -> {
-                anInt2987 = buffer.readUnsignedByte()
-                anInt3008 = buffer.readShort()
-            }
-            100 -> {
-                anInt3038 = buffer.readUnsignedByte()
-                anInt3013 = buffer.readShort()
-            }
-            101 -> anInt2958 = buffer.readUnsignedByte()
-            102 -> mapscene = buffer.readShort()
-            103 -> culling = 0
-            104 -> anInt3024 = buffer.readUnsignedByte()
-            105 -> invertMapScene = true
-            106 -> {
-                val length = buffer.readUnsignedByte()
-                var total = 0
-                animations = IntArray(length)
-                percents = IntArray(length)
-                for (count in 0 until length) {
-                    animations!![count] = buffer.readShort()
-                    if (animations!![count] == 65535) {
-                        animations!![count] = -1
-                    }
-                    percents!![count] = buffer.readUnsignedByte()
-                    total += percents!![count]
-                }
-                for (count in 0 until length) {
-                    percents!![count] = 65535 * percents!![count] / total
-                }
-            }
-            107 -> mapDefinitionId = buffer.readShort()
+            106 -> buffer.skip(buffer.readUnsignedByte() * 3)
             in 150..154 -> {
                 if (options == null) {
                     options = arrayOf(null, null, null, null, null, "Examine")
@@ -154,31 +75,9 @@ open class ObjectDecoder(
                     options!![opcode - 150] = null
                 }
             }
-            160 -> anIntArray2981 = IntArray(buffer.readUnsignedByte()) { buffer.readShort() }
-            162 -> {
-                contouredGround = 3
-                anInt3023 = buffer.readInt()
-            }
-            163 -> {
-                aByte2974 = buffer.readByte().toByte()
-                aByte3045 = buffer.readByte().toByte()
-                aByte3052 = buffer.readByte().toByte()
-                aByte2960 = buffer.readByte().toByte()
-            }
-            164 -> anInt2964 = buffer.readUnsignedShort()
-            165 -> anInt2963 = buffer.readUnsignedShort()
-            166 -> anInt3018 = buffer.readUnsignedShort()
-            167 -> anInt2983 = buffer.readShort()
-            168 -> aBoolean2961 = true
-            169 -> aBoolean2993 = true
-            170 -> anInt3032 = buffer.readSmart()
-            171 -> anInt2962 = buffer.readSmart()
-            173 -> {
-                anInt3050 = buffer.readShort()
-                anInt3020 = buffer.readShort()
-            }
-            177 -> aBoolean2992 = true
-            178 -> anInt2975 = buffer.readUnsignedByte()
+            160 -> buffer.skip(buffer.readUnsignedByte() * 2)
+            162, 163, 173 -> buffer.skip(4)
+            170, 171 -> buffer.readSmart()
             249 -> readParameters(buffer)
         }
     }
