+++ tools/src/main/kotlin/world/gregs/voidps/tools/AnimationDefinitions.kt
@@ -2,13 +2,13 @@ package world.gregs.voidps.tools
 
 import world.gregs.voidps.cache.Cache
 import world.gregs.voidps.cache.CacheDelegate
-import world.gregs.voidps.cache.definition.decoder.AnimationDecoder
+import world.gregs.voidps.cache.definition.decoder.AnimationDecoderFull
 
 object AnimationDefinitions {
     @JvmStatic
     fun main(args: Array<String>) {
         val cache: Cache = CacheDelegate(property("cachePath"))
-        val decoder = AnimationDecoder().loadCache(cache)
+        val decoder = AnimationDecoderFull().loadCache(cache)
         loop@ for (i in decoder.indices) {
             val def = decoder.getOrNull(i) ?: continue
             if (def.aBoolean691) {
